requirements.txt
Flask==2.0.1
jsonpickle==2.0.0
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
scipy==1.7.0
torch==1.9.0
tqdm==4.61.2
